+++ lib/nretnil-utilities.rb
@@ -1,4 +1,5 @@
 require 'nretnil/utilities/array'
 require 'nretnil/utilities/hash'
+require 'nretnil/utilities/integer'
 require 'nretnil/utilities/string'
 require 'nretnil/utilities/version'