+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+<h4>0.0.2</h4>
+
+* Added Integer.to_human for file sizes
+
 <h4>0.0.1</h4>
 
 * Initial Gem Build